+++ src/main/scala/com/raquo/laminar/modifiers/ChildrenCommandInserter.scala
@@ -4,7 +4,7 @@ import com.raquo.airstream.eventstream.EventStream
 import com.raquo.laminar.CollectionCommand
 import com.raquo.laminar.lifecycle.{InsertContext, MountContext}
 import com.raquo.laminar.modifiers.ChildrenInserter.Child
-import com.raquo.laminar.nodes.{ChildNode, ReactiveElement}
+import com.raquo.laminar.nodes.{ChildNode, ParentNode, ReactiveElement}
 
 import scala.scalajs.js
 
@@ -42,50 +42,51 @@ object ChildrenCommandInserter {
     var nodeCountDiff = 0
     command match {
       case CollectionCommand.Append(node) =>
-        val sentinelIndex = parentNode.indexOfChild(sentinelNode)
-        if (parentNode.insertChild(node, atIndex = sentinelIndex + extraNodeCount + 1)) {
+        val sentinelIndex = ParentNode.indexOfChild(parent = parentNode, sentinelNode)
+        if (ParentNode.insertChild(parent = parentNode, child = node, atIndex = sentinelIndex + extraNodeCount + 1)) {
           nodeCountDiff = 1
         }
         nodeCountDiff = 1
       case CollectionCommand.Prepend(node) =>
-        val sentinelIndex = parentNode.indexOfChild(sentinelNode)
-        if (parentNode.insertChild(node, atIndex = sentinelIndex + 1)) {
+        val sentinelIndex = ParentNode.indexOfChild(parent = parentNode, sentinelNode)
+        if (ParentNode.insertChild(parent = parentNode, child = node, atIndex = sentinelIndex + 1)) {
           nodeCountDiff = 1
         }
       case CollectionCommand.Insert(node, atIndex) =>
-        val sentinelIndex = parentNode.indexOfChild(sentinelNode)
-        if (parentNode.insertChild(node, atIndex = sentinelIndex + atIndex + 1)) {
+        val sentinelIndex = ParentNode.indexOfChild(parent = parentNode, sentinelNode)
+        if (ParentNode.insertChild(parent = parentNode, child = node, atIndex = sentinelIndex + atIndex + 1)) {
           nodeCountDiff = 1
         }
       case CollectionCommand.Remove(node) =>
-        if (parentNode.removeChild(node)) {
+        if (ParentNode.removeChild(parent = parentNode, child = node)) {
           nodeCountDiff = -1
         }
       case CollectionCommand.Move(node, toIndex) =>
         // @TODO same as Insert.
         // @TODO Should we also add a MoveToEnd method?
-        val sentinelIndex = parentNode.indexOfChild(sentinelNode)
-        if (parentNode.insertChild(node, atIndex = sentinelIndex + toIndex + 1)) {
+        val sentinelIndex = ParentNode.indexOfChild(parent = parentNode, sentinelNode)
+        if (ParentNode.insertChild(parent = parentNode, child = node, atIndex = sentinelIndex + toIndex + 1)) {
           nodeCountDiff = 1
         }
       case CollectionCommand.Replace(node, withNode) =>
-        parentNode.replaceChild(oldChild = node, newChild = withNode)
+        ParentNode.replaceChild(parent = parentNode, oldChild = node, newChild = withNode)
       case CollectionCommand.ReplaceAll(newNodes) =>
-        val sentinelIndex = parentNode.indexOfChild(sentinelNode)
+        val sentinelIndex = ParentNode.indexOfChild(parent = parentNode, sentinelNode)
         if (extraNodeCount == 0) {
           var numInsertedNodes = 0
           newNodes.foreach { newChild =>
-            if (parentNode.insertChild(newChild, atIndex = sentinelIndex + 1 + numInsertedNodes)) {
+            if (ParentNode.insertChild(parent = parentNode, child = newChild, atIndex = sentinelIndex + 1 + numInsertedNodes)) {
               numInsertedNodes += 1
             }
           }
           nodeCountDiff = numInsertedNodes
         } else {
           val oldNodeCount = extraNodeCount
-          val replaced = parentNode.replaceChildren(
+          val replaced = ParentNode.replaceChildren(
+            parent = parentNode,
             fromIndex = sentinelIndex + 1,
             toIndex = sentinelIndex + extraNodeCount,
-            js.Array(newNodes.toSeq: _*) // @TODO[Performance]
+            newNodes
           )
           if (replaced) {
             nodeCountDiff = newNodes.size - oldNodeCount