+++ src/main/scala/com/raquo/laminar/modifiers/ChildrenInserter.scala
@@ -4,7 +4,7 @@ import com.raquo.airstream.core.Observable
 import com.raquo.airstream.eventstream.EventStream
 import com.raquo.airstream.signal.Signal
 import com.raquo.laminar.lifecycle.{InsertContext, MountContext}
-import com.raquo.laminar.nodes.{ChildNode, ReactiveElement}
+import com.raquo.laminar.nodes.{ChildNode, ParentNode, ReactiveElement}
 import org.scalajs.dom
 
 import scala.collection.immutable
@@ -51,7 +51,7 @@ object ChildrenInserter {
     prevChildrenCount: Int
   ): Int = {
     val liveNodeList = parentNode.ref.childNodes
-    val sentinelIndex = parentNode.indexOfChild(sentinelNode)
+    val sentinelIndex = ParentNode.indexOfChild(parent = parentNode, sentinelNode)
 
     // Loop variables
     var index = 0
@@ -84,7 +84,7 @@ object ChildrenInserter {
         // Note: `prevChildRef` is not valid in this branch
         // println("> overflow: inserting " + nextChild.ref.textContent + " at index " + nextChildNodeIndex)
         // @Note: DOM update
-        parentNode.insertChild(nextChild, atIndex = nextChildNodeIndex)
+        ParentNode.insertChild(parent = parentNode, child = nextChild, atIndex = nextChildNodeIndex)
         prevChildRef = nextChild.ref
         currentChildrenCount += 1
       } else {
@@ -98,7 +98,7 @@ object ChildrenInserter {
             // nextChild not found in prevChildren, so it's a new child, so we need to insert it
             // println("> new: inserting " + nextChild.ref.textContent + " at index " + nextChildNodeIndex)
             // @Note: DOM update
-            parentNode.insertChild(nextChild, atIndex = nextChildNodeIndex)
+            ParentNode.insertChild(parent = parentNode, child = nextChild, atIndex = nextChildNodeIndex)
             prevChildRef = nextChild.ref
             currentChildrenCount += 1
           } else {
@@ -119,7 +119,7 @@ object ChildrenInserter {
               val prevChild = prevChildFromRef(prevChildren, prevChildRef)
               // println("> removing " + prevChild.ref.textContent)
               // @Note: DOM update
-              parentNode.removeChild(prevChild)
+              ParentNode.removeChild(parent = parentNode, child = prevChild)
               prevChildRef = nextPrevChildRef
               currentChildrenCount -= 1
             }
@@ -127,7 +127,7 @@ object ChildrenInserter {
               // nextChild is still not in the right place, so let's move it to the correct index
               // println("> order: inserting " + nextChild.ref.textContent + " at index " + nextChildNodeIndex)
               // @Note: DOM update
-              parentNode.insertChild(nextChild, atIndex = nextChildNodeIndex)
+              ParentNode.insertChild(parent = parentNode, child = nextChild, atIndex = nextChildNodeIndex)
               prevChildRef = nextChild.ref
               // This is a MOVE, so we DO NOT update currentDomChildrenCount here.
             }
@@ -144,7 +144,7 @@ object ChildrenInserter {
       val nextPrevChildRef = prevChildRef.nextSibling
       // Whenever we insert, move or remove items from the DOM, we need to manually update `prevChildRef` to point to the node at the current index
       // @Note: DOM update
-      parentNode.removeChild(prevChildFromRef(prevChildren, prevChildRef))
+      ParentNode.removeChild(parent = parentNode, child = prevChildFromRef(prevChildren, prevChildRef))
       prevChildRef = nextPrevChildRef
       currentChildrenCount -= 1
     }